front-end/api/requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
google-generativeai>=0.3.0
PyPDF2>=2.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.0.2
numpy>=1.21.0
